refactor(frontend): mirror backend Auth/Users modules in api layer
Replace src/api/account with api/auth (session, registration, email
confirmation, anonymous password reset) and api/users/CurrentUserApiHooks
(profile, password change, publications, cruise effects), matching the
backend's Auth and Users/Me vertical slices. The shared ProblemDetails
type moves to lib/api next to the axios client. Hook names are
unchanged; consumers only needed import-path updates.
