+++ frontend/src/api/auth/contracts.ts
@@ -35,42 +35,9 @@ export type ResetPasswordRequest = {
   passwordConfirm: string;
 };
 
-export type ChangePasswordRequest = {
-  password: string;
-  newPassword: string;
-};
-
-export type CurrentPublicationResponse = {
-  id: string;
-  category: string;
-  doi: string | null;
-  authors: string | null;
-  title: string | null;
-  magazine: string | null;
-  year: string | null;
-  ministerialPoints: string;
-};
-
-export type CurrentPublicationImportRequest = Omit<CurrentPublicationResponse, 'id'>;
-
 export type AuthResponse = {
   accessToken: string;
   accessTokenExpirationDate: string;
   refreshToken: string;
   refreshTokenExpirationDate: string;
 };
-
-export type CurrentUserResponse = {
-  id: string;
-  email: string;
-  firstName: string;
-  lastName: string;
-  roles: Role[];
-  emailConfirmed: boolean;
-  accepted: boolean;
-};
-
-export type ProblemDetails = {
-  detail?: string;
-};
-import { Role } from '@/models/shared/Role';
