+++ frontend/src/api/users/contracts.ts
@@ -29,3 +29,31 @@ export type CruiseManagerOptionResponse = {
   firstName: string;
   lastName: string;
 };
+
+export type CurrentUserResponse = {
+  id: string;
+  email: string;
+  firstName: string;
+  lastName: string;
+  roles: Role[];
+  emailConfirmed: boolean;
+  accepted: boolean;
+};
+
+export type ChangePasswordRequest = {
+  password: string;
+  newPassword: string;
+};
+
+export type CurrentPublicationResponse = {
+  id: string;
+  category: string;
+  doi: string | null;
+  authors: string | null;
+  title: string | null;
+  magazine: string | null;
+  year: string | null;
+  ministerialPoints: string;
+};
+
+export type CurrentPublicationImportRequest = Omit<CurrentPublicationResponse, 'id'>;
