+++ frontend/src/providers/UserContextProvider.tsx
@@ -6,7 +6,8 @@ import { client, setAuthToken } from '@/lib/api';
 import { setSentryUser } from '@/lib/sentry';
 import { Role } from '@/models/shared/Role';
 import { UserContext, UserContextType } from '@/providers/UserContext';
-import { useLoginMutation, useProfileQuery, useRefreshTokenMutation } from '@/api/account/AccountAuthApiHooks';
+import { useLoginMutation, useRefreshTokenMutation } from '@/api/auth/AuthApiHooks';
+import { useProfileQuery } from '@/api/users/CurrentUserApiHooks';
 import { AuthDetails } from '@/models/user/AuthDetails';
 import { SignInResult } from '@/models/user/Results';
 import { getStoredAuthDetails, setStoredAuthDetails } from '@/providers/StoredAuthDetails';
