+++ frontend/src/routes/(auth)/confirm-email.tsx
@@ -8,7 +8,7 @@ import { AppButton } from '@/components/shared/AppButton';
 import { AppLayout } from '@/components/shared/AppLayout';
 import { AppLink } from '@/components/shared/AppLink';
 import { AppLoader } from '@/components/shared/layout/AppLoader';
-import { useConfirmEmailMutation } from '@/api/account/AccountRecoveryApiHooks';
+import { useConfirmEmailMutation } from '@/api/auth/AuthApiHooks';
 import { Result } from '@/models/user/Results';
 
 export const Route = createFileRoute('/(auth)/confirm-email')({
