+++ frontend/src/routes/(auth)/forgot-password.tsx
@@ -9,7 +9,7 @@ import { AppLink } from '@/components/shared/AppLink';
 import { AppFloatingLabelInput } from '@/components/shared/inputs/AppFloatingLabelInput';
 import { trackFormSubmit } from '@/lib/sentry';
 import { getErrors } from '@/lib/utils';
-import { useForgotPasswordMutation } from '@/api/account/AccountRecoveryApiHooks';
+import { useForgotPasswordMutation } from '@/api/auth/AuthApiHooks';
 import { Result } from '@/models/user/Results';
 
 export const Route = createFileRoute('/(auth)/forgot-password')({
